+++ alembic/env.py
@@ -5,7 +5,7 @@ from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
 from server.base.models import Base
-from server.handlers import engine_url
+from server.sources import engine_url
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
