+++ config.example.yaml
@@ -1,13 +1,14 @@
-TOKEN:
-WEBHOOK_URL:
-SECRET_TOKEN:
+TGBOT: # True or False (if True, the bot will be used)
+TOKEN: # required if TGBOT is True
+WEBHOOK_URL: # required if TGBOT is True
+SECRET_TOKEN: # required if TGBOT is True
 DEV: # True or False
 PGUSER:
 PGPASSWORD:
 PGPORT:
 PGHOST:
 PGDATABASE:
-ADMIN_TG_ID: # Telegram user ID of the admin
+ADMIN_TG_ID: # Telegram user ID of the admin, required if TGBOT is True
 SSL_KEYFILE: # Path to the SSL key file
 SSL_CERTFILE: # Path to the SSL certificate file
 TYPESENSE_API_KEY:
